add reflect native for prover, make some functions pub, add pontem/bech32 addresses parsing

+++ language/tools/move-package/src/compilation/build_plan.rs
@@ -99,7 +99,16 @@ impl BuildPlan {
         for dir in std::fs::read_dir(build_root)? {
             let path = dir?.path();
             if !keep_paths.iter().any(|name| path.ends_with(name.as_str())) {
-                std::fs::remove_dir_all(&path)?;
+                // additional handling of files is required for Dove, which uses special
+                // .version file inside the ./build directory to be able to detect `Move.toml`
+                // changes between invocations
+                if path.exists() {
+                    if path.is_dir() {
+                        std::fs::remove_dir_all(&path)?;
+                    } else {
+                        std::fs::remove_file(&path)?;
+                    }
+                }
             }
         }
         Ok(())
